requirements.txt
# Django and Core Dependencies
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
django-debug-toolbar==4.2.0
django-widget-tweaks==1.5.0
django-crispy-forms==2.3
crispy-bootstrap5==2024.10
django-phonenumber-field==7.3.0
django-chartjs==2.3.0
drf-yasg==1.21.7
django-extensions==3.2.3
django-sslserver==0.22
django-secure==1.0.2
django-storages==1.14.2
django-redis==5.4.0

# Database and Cache
psycopg2-binary==2.9.9
redis==5.0.1

# Celery
celery==5.3.6

# Development and Testing
black==24.2.0
flake8==7.0.0
pytest==8.0.2
pytest-django==4.8.0
coverage==7.3.2
Werkzeug==3.0.1

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Security and SSL
pyOpenSSL==24.0.0
certifi==2024.2.2
cryptography==42.0.2
urllib3==2.2.1

# Utilities
requests==2.31.0
python-dotenv==1.0.1
qrcode==7.4.2
Pillow==10.2.0
reportlab==4.0.8
xlsxwriter==3.1.9
phonenumbers==8.13.27
boto3==1.34.14
twilio==8.12.0
python-magic==0.4.27 